refactor: pass plugin-ai-agents types to core and import them in plugins that use it
