+++ packages/botonic-core/src/models/index.ts
@@ -1,3 +1,4 @@
+export * from './ai-agents'
 export * from './events'
 export * from './hubtype-analytics'
 export * from './knowledge-bases'
