+++ packages/botonic-plugin-ai-agents/src/structured-output/index.ts
@@ -1,18 +1,10 @@
+import { OutputMessage } from '@botonic/core'
 import { z } from 'zod'
 
-import { CarouselMessage, CarouselSchema } from './carousel'
-import { ExitMessage, ExitSchema } from './exit'
-import { TextMessage, TextSchema } from './text'
-import {
-  TextWithButtonsMessage,
-  TextWithButtonsSchema,
-} from './text-with-buttons'
-
-export type OutputMessage =
-  | TextMessage
-  | TextWithButtonsMessage
-  | CarouselMessage
-  | ExitMessage
+import { CarouselSchema } from './carousel'
+import { ExitSchema } from './exit'
+import { TextSchema } from './text'
+import { TextWithButtonsSchema } from './text-with-buttons'
 
 export interface Output {
   messages: OutputMessage[]
