--- packages/botonic-plugin-ai-agents/src/structured-output/shared.ts
@@ -1,3 +1,0 @@
-export interface BaseMessage {
-  type: 'text' | 'textWithButtons' | 'carousel' | 'exit'
-}
