+++ packages/botonic-plugin-flow-builder/src/types.ts
@@ -1,8 +1,13 @@
 import {
+  AgenticOutputMessage,
   BotContext,
+  CarouselMessage,
+  InferenceResponse,
   KnowledgeBaseResponse,
   PluginPreRequest,
   ResolvedPlugins,
+  TextMessage,
+  TextWithButtonsMessage,
 } from '@botonic/core'
 
 import { FlowContent } from './content-fields'
@@ -104,48 +109,4 @@ export interface SmartIntentResponse {
 
 export interface PayloadParamsBase {
   followUpContentID?: string
-}
-
-export interface OutputBaseMessage {
-  type: 'text' | 'textWithButtons' | 'carousel' | 'exit'
-}
-
-export interface TextMessage extends OutputBaseMessage {
-  type: 'text'
-  content: {
-    text: string
-  }
-}
-
-export interface TextWithButtonsMessage extends OutputBaseMessage {
-  type: 'textWithButtons'
-  content: {
-    text: string
-    buttons: string[]
-  }
-}
-
-export interface CarouselMessage extends OutputBaseMessage {
-  type: 'carousel'
-  content: {
-    elements: {
-      title: string
-      subtitle: string
-      image: string
-      button: { text: string; url: string }
-    }[]
-  }
-}
-
-export type AgenticOutputMessage =
-  | TextMessage
-  | TextWithButtonsMessage
-  | CarouselMessage
-
-export interface InferenceResponse {
-  messages: AgenticOutputMessage[]
-  toolsExecuted: string[]
-  exit: boolean
-  inputGuardrailTriggered: boolean
-  outputGuardrailTriggered: boolean
 }
