+++ packages/botonic-plugin-flow-builder/tests/__mocks__/ai-agent.ts
@@ -1,4 +1,4 @@
-import { AgenticOutputMessage, InferenceResponse } from '../../src/types'
+import { AgenticOutputMessage, InferenceResponse } from '@botonic/core'
 
 export function mockAiAgentResponse(messages: AgenticOutputMessage[]) {
   return jest.fn(() => {
